+++ src/main/java/org/openlmis/stockmanagement/repository/CalculatedStockOnHandRepository.java
@@ -15,11 +15,27 @@
 
 package org.openlmis.stockmanagement.repository;
 
+import java.time.LocalDate;
+import java.util.Collection;
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import org.openlmis.stockmanagement.domain.event.CalculatedStockOnHand;
 import org.springframework.data.jpa.repository.JpaRepository;
-
+import org.springframework.data.repository.query.Param;
 
 public interface CalculatedStockOnHandRepository
-        extends JpaRepository<CalculatedStockOnHand, UUID> {
+    extends JpaRepository<CalculatedStockOnHand, UUID> {
+
+  Optional<CalculatedStockOnHand>
+      findFirstByStockCardIdAndOccurredDateLessThanEqualOrderByOccurredDateDesc(
+          @Param("stockCardId") UUID stockCardId,
+          @Param("asOfDate") LocalDate asOfDate);
+
+  List<CalculatedStockOnHand>
+      findByStockCardIdAndOccurredDateGreaterThanEqualOrderByOccurredDateAsc(
+          UUID stockCardId, LocalDate asOfDate);
+
+  List<CalculatedStockOnHand> findByStockCardIdAndOccurredDateBetween(
+      Collection<UUID> stockCardId, LocalDate startDate, LocalDate endDate);
 }