+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 New functionality:
 * [OLMIS-6368](https://openlmis.atlassian.net/browse/OLMIS-6368): Create SoH fact table.
 * [OLMIS-6434](https://openlmis.atlassian.net/browse/OLMIS-6434): Calculated and populated Stock on Hand values to a new table.
+* [OLMIS-6483](https://openlmis.atlassian.net/browse/OLMIS-6483): Added saving calculated SoH value while sensing stock event and using it during getting stock card info.
 * [OLMIS-6558](https://openlmis.atlassian.net/browse/OLMIS-6558): Add new environment variable - PUBLIC_URL and use to for email generated links
+
 Improvements:
 * [OLMIS-5569](https://openlmis.atlassian.net/browse/OLMIS-5569): Moved reason TRANSFER_IN from demo data to bootstrap data.
 * [OLMIS-6408](https://openlmis.atlassian.net/browse/OLMIS-6408): Added pageable validator.
+* [OLMIS-6474](https://openlmis.atlassian.net/browse/OLMIS-6474): Performance improvements of `GET /api/orderableFulfills` endpoint.
 
 4.1.1 / 2019-05-27
 ==================
@@ -17,6 +20,7 @@ Improvements:
 
 Bug fixes, security and performance improvements, also backwards-compatible:
 * [OLMIS-6304](https://openlmis.atlassian.net/browse/OLMIS-6304): Fixed displaying product names in reports.
+* [OLMIS-6531](https://openlmis.atlassian.net/browse/OLMIS-6531): Fix: only the newest version of a product is visible in the Physical Inventory printout.
 
 4.1.0 / 2019-05-27
 ==================